Deploy artifacts to dev.gpkpw.com, under a folder for the branch
Every successful build (both master and off-master builds)
results in an upload of build artifacts to a gcs bucket that's
an http static website, to a subfolder on that site that's
named after the git branch.

This allows us to easily share off-master work with each
other.

basic script that uploads all files in a dir to gcs
integration deploy to gcs into the build